+++ exporter/opentelemetry-exporter-jaeger/src/opentelemetry/exporter/jaeger/__init__.py
@@ -15,9 +15,11 @@
 
 """
 The **OpenTelemetry Jaeger Exporter** allows to export `OpenTelemetry`_ traces to `Jaeger`_.
-This exporter always send traces to the configured agent using Thrift compact protocol over UDP.
-An optional collector can be configured, in this case Thrift binary protocol over HTTP is used.
-gRPC is still not supported by this implementation.
+This exporter always sends traces to the configured agent using the Thrift compact protocol over UDP.
+When it is not feasible to deploy Jaeger Agent next to the application, for example, when the
+application code is running as Lambda function, a collector can be configured to send spans
+using either Thrift over HTTP or Protobuf via gRPC. If both agent and collector are configured,
+the exporter sends traces only to the collector to eliminate the duplicate entries.
 
 Usage
 -----
@@ -42,6 +44,9 @@ Usage
         # collector_endpoint='http://localhost:14268/api/traces?format=jaeger.thrift',
         # username=xxxx, # optional
         # password=xxxx, # optional
+        # insecure=True, # optional
+        # credentials=xxx # optional channel creds
+        # transport_format='protobuf' # optional
     )
 
     # Create a BatchExportSpanProcessor and add the exporter to it
@@ -58,33 +63,47 @@ API
 .. _Jaeger: https://www.jaegertracing.io/
 .. _OpenTelemetry: https://github.com/open-telemetry/opentelemetry-python/
 """
+# pylint: disable=protected-access
 
 import base64
 import logging
 import socket
+from typing import Optional, Union
 
+from grpc import (
+    ChannelCredentials,
+    insecure_channel,
+    secure_channel,
+    ssl_channel_credentials,
+)
 from thrift.protocol import TBinaryProtocol, TCompactProtocol
 from thrift.transport import THttpClient, TTransport
 
 from opentelemetry.configuration import Configuration
+from opentelemetry.exporter.jaeger import util
+from opentelemetry.exporter.jaeger.gen import model_pb2
 from opentelemetry.exporter.jaeger.gen.agent import Agent as agent
-from opentelemetry.exporter.jaeger.gen.jaeger import Collector as jaeger
+from opentelemetry.exporter.jaeger.gen.collector_pb2 import PostSpansRequest
+from opentelemetry.exporter.jaeger.gen.collector_pb2_grpc import (
+    CollectorServiceStub,
+)
+from opentelemetry.exporter.jaeger.gen.jaeger import Collector as jaeger_thrift
+from opentelemetry.exporter.jaeger.send.thrift import AgentClientUDP, Collector
+from opentelemetry.exporter.jaeger.translate import Translate
+from opentelemetry.exporter.jaeger.translate.protobuf import ProtobufTranslator
+from opentelemetry.exporter.jaeger.translate.thrift import ThriftTranslator
 from opentelemetry.sdk.trace.export import Span, SpanExporter, SpanExportResult
 from opentelemetry.trace import SpanKind
 from opentelemetry.trace.status import StatusCode
 
 DEFAULT_AGENT_HOST_NAME = "localhost"
 DEFAULT_AGENT_PORT = 6831
+DEFAULT_GRPC_COLLECTOR_ENDPOINT = "localhost:14250"
 
 UDP_PACKET_MAX_LENGTH = 65000
 
-OTLP_JAEGER_SPAN_KIND = {
-    SpanKind.CLIENT: "client",
-    SpanKind.SERVER: "server",
-    SpanKind.CONSUMER: "consumer",
-    SpanKind.PRODUCER: "producer",
-    SpanKind.INTERNAL: "internal",
-}
+TRANSPORT_FORMAT_THRIFT = "thrift"
+TRANSPORT_FORMAT_PROTOBUF = "protobuf"
 
 logger = logging.getLogger(__name__)
 
@@ -97,21 +116,28 @@ class JaegerSpanExporter(SpanExporter):
             when query for spans.
         agent_host_name: The host name of the Jaeger-Agent.
         agent_port: The port of the Jaeger-Agent.
-        collector_endpoint: The endpoint of the Jaeger-Collector HTTP/HTTPS Thrift.
+        collector_endpoint: The endpoint of the Jaeger collector that uses
+            Thrift over HTTP/HTTPS or Protobuf via gRPC.
         username: The user name of the Basic Auth if authentication is
             required.
         password: The password of the Basic Auth if authentication is
             required.
+        insecure: True if collector has no encryption or authentication
+        credentials: Credentials for server authentication.
+        transport_format: Transport format for exporting spans to collector.
     """
 
     def __init__(
         self,
-        service_name,
-        agent_host_name=None,
-        agent_port=None,
-        collector_endpoint=None,
-        username=None,
-        password=None,
+        service_name: str,
+        agent_host_name: Optional[str] = None,
+        agent_port: Optional[int] = None,
+        collector_endpoint: Optional[str] = None,
+        username: Optional[str] = None,
+        password: Optional[str] = None,
+        insecure: Optional[bool] = None,
+        credentials: Optional[ChannelCredentials] = None,
+        transport_format: Optional[str] = None,
     ):
         self.service_name = service_name
         self.agent_host_name = _parameter_setter(
@@ -124,7 +150,9 @@ class JaegerSpanExporter(SpanExporter):
             env_variable=Configuration().EXPORTER_JAEGER_AGENT_PORT,
             default=DEFAULT_AGENT_PORT,
         )
-        self._agent_client = None
+        self._agent_client = AgentClientUDP(
+            host_name=self.agent_host_name, port=self.agent_port
+        )
         self.collector_endpoint = _parameter_setter(
             param=collector_endpoint,
             env_variable=Configuration().EXPORTER_JAEGER_ENDPOINT,
@@ -141,21 +169,42 @@ class JaegerSpanExporter(SpanExporter):
             default=None,
         )
         self._collector = None
+        self._grpc_client = None
+        self.insecure = util._get_insecure(insecure)
+        self.credentials = util._get_credentials(credentials)
+        self.transport_format = (
+            transport_format.lower()
+            if transport_format
+            else TRANSPORT_FORMAT_THRIFT
+        )
 
     @property
-    def agent_client(self):
-        if self._agent_client is None:
-            self._agent_client = AgentClientUDP(
-                host_name=self.agent_host_name, port=self.agent_port
-            )
-        return self._agent_client
+    def _collector_grpc_client(self) -> Optional[CollectorServiceStub]:
+        if self.transport_format != TRANSPORT_FORMAT_PROTOBUF:
+            return None
+
+        endpoint = self.collector_endpoint or DEFAULT_GRPC_COLLECTOR_ENDPOINT
+
+        if self._grpc_client is None:
+            if self.insecure:
+                self._grpc_client = CollectorServiceStub(
+                    insecure_channel(endpoint)
+                )
+            else:
+                self._grpc_client = CollectorServiceStub(
+                    secure_channel(endpoint, self.credentials)
+                )
+        return self._grpc_client
 
     @property
-    def collector(self):
+    def _collector_http_client(self) -> Optional[Collector]:
         if self._collector is not None:
             return self._collector
 
-        if self.collector_endpoint is None:
+        if (
+            self.collector_endpoint is None
+            or self.transport_format != TRANSPORT_FORMAT_THRIFT
+        ):
             return None
 
         auth = None
@@ -167,18 +216,25 @@ class JaegerSpanExporter(SpanExporter):
         )
         return self._collector
 
-    def export(self, spans):
-        jaeger_spans = _translate_to_jaeger(spans)
-
-        batch = jaeger.Batch(
-            spans=jaeger_spans,
-            process=jaeger.Process(serviceName=self.service_name),
-        )
-
-        if self.collector is not None:
-            self.collector.submit(batch)
+    def export(self, spans) -> SpanExportResult:
+        translator = Translate(spans)
+        if self.transport_format == TRANSPORT_FORMAT_PROTOBUF:
+            pb_translator = ProtobufTranslator(self.service_name)
+            jaeger_spans = translator._translate(pb_translator)
+            batch = model_pb2.Batch(spans=jaeger_spans)
+            request = PostSpansRequest(batch=batch)
+            self._collector_grpc_client.PostSpans(request)
         else:
-            self.agent_client.emit(batch)
+            thrift_translator = ThriftTranslator()
+            jaeger_spans = translator._translate(thrift_translator)
+            batch = jaeger_thrift.Batch(
+                spans=jaeger_spans,
+                process=jaeger_thrift.Process(serviceName=self.service_name),
+            )
+            if self._collector_http_client is not None:
+                self._collector_http_client.submit(batch)
+            else:
+                self._agent_client.emit(batch)
 
         return SpanExportResult.SUCCESS
 
@@ -200,269 +256,3 @@ def _parameter_setter(param, env_variable, default):
         res = param
 
     return res
-
-
-def _nsec_to_usec_round(nsec):
-    """Round nanoseconds to microseconds"""
-    return (nsec + 500) // 10 ** 3
-
-
-def _translate_to_jaeger(spans: Span):
-    """Translate the spans to Jaeger format.
-
-    Args:
-        spans: Tuple of spans to convert
-    """
-
-    jaeger_spans = []
-
-    for span in spans:
-        ctx = span.get_span_context()
-        trace_id = ctx.trace_id
-        span_id = ctx.span_id
-
-        start_time_us = _nsec_to_usec_round(span.start_time)
-        duration_us = _nsec_to_usec_round(span.end_time - span.start_time)
-
-        status = span.status
-
-        parent_id = span.parent.span_id if span.parent else 0
-
-        tags = _extract_tags(span.attributes)
-        tags.extend(_extract_tags(span.resource.attributes))
-
-        tags.extend(
-            [
-                _get_long_tag("status.code", status.status_code.value),
-                _get_string_tag("status.message", status.description),
-                _get_string_tag("span.kind", OTLP_JAEGER_SPAN_KIND[span.kind]),
-            ]
-        )
-
-        if span.instrumentation_info is not None:
-            tags.extend(
-                [
-                    _get_string_tag(
-                        "otel.instrumentation_library.name",
-                        span.instrumentation_info.name,
-                    ),
-                    _get_string_tag(
-                        "otel.instrumentation_library.version",
-                        span.instrumentation_info.version,
-                    ),
-                ]
-            )
-
-        # Ensure that if Status.Code is not OK, that we set the "error" tag on the Jaeger span.
-        if not status.is_ok:
-            tags.append(_get_bool_tag("error", True))
-
-        refs = _extract_refs_from_span(span)
-        logs = _extract_logs_from_span(span)
-
-        flags = int(ctx.trace_flags)
-
-        jaeger_span = jaeger.Span(
-            traceIdHigh=_get_trace_id_high(trace_id),
-            traceIdLow=_get_trace_id_low(trace_id),
-            # generated code expects i64
-            spanId=_convert_int_to_i64(span_id),
-            operationName=span.name,
-            startTime=start_time_us,
-            duration=duration_us,
-            tags=tags,
-            logs=logs,
-            references=refs,
-            flags=flags,
-            parentSpanId=_convert_int_to_i64(parent_id),
-        )
-
-        jaeger_spans.append(jaeger_span)
-
-    return jaeger_spans
-
-
-def _extract_refs_from_span(span):
-    if not span.links:
-        return None
-
-    refs = []
-    for link in span.links:
-        trace_id = link.context.trace_id
-        span_id = link.context.span_id
-        refs.append(
-            jaeger.SpanRef(
-                refType=jaeger.SpanRefType.FOLLOWS_FROM,
-                traceIdHigh=_get_trace_id_high(trace_id),
-                traceIdLow=_get_trace_id_low(trace_id),
-                spanId=_convert_int_to_i64(span_id),
-            )
-        )
-    return refs
-
-
-def _convert_int_to_i64(val):
-    """Convert integer to signed int64 (i64)"""
-    if val > 0x7FFFFFFFFFFFFFFF:
-        val -= 0x10000000000000000
-    return val
-
-
-def _get_trace_id_low(trace_id):
-    return _convert_int_to_i64(trace_id & 0xFFFFFFFFFFFFFFFF)
-
-
-def _get_trace_id_high(trace_id):
-    return _convert_int_to_i64((trace_id >> 64) & 0xFFFFFFFFFFFFFFFF)
-
-
-def _extract_logs_from_span(span):
-    if not span.events:
-        return None
-
-    logs = []
-
-    for event in span.events:
-        fields = _extract_tags(event.attributes)
-
-        fields.append(
-            jaeger.Tag(
-                key="message", vType=jaeger.TagType.STRING, vStr=event.name
-            )
-        )
-
-        event_timestamp_us = _nsec_to_usec_round(event.timestamp)
-        logs.append(
-            jaeger.Log(timestamp=int(event_timestamp_us), fields=fields)
-        )
-    return logs
-
-
-def _extract_tags(attr):
-    if not attr:
-        return []
-    tags = []
-    for attribute_key, attribute_value in attr.items():
-        tag = _convert_attribute_to_tag(attribute_key, attribute_value)
-        if tag is None:
-            continue
-        tags.append(tag)
-    return tags
-
-
-def _convert_attribute_to_tag(key, attr):
-    """Convert the attributes to jaeger tags."""
-    if isinstance(attr, bool):
-        return jaeger.Tag(key=key, vBool=attr, vType=jaeger.TagType.BOOL)
-    if isinstance(attr, str):
-        return jaeger.Tag(key=key, vStr=attr, vType=jaeger.TagType.STRING)
-    if isinstance(attr, int):
-        return jaeger.Tag(key=key, vLong=attr, vType=jaeger.TagType.LONG)
-    if isinstance(attr, float):
-        return jaeger.Tag(key=key, vDouble=attr, vType=jaeger.TagType.DOUBLE)
-    if isinstance(attr, tuple):
-        return jaeger.Tag(key=key, vStr=str(attr), vType=jaeger.TagType.STRING)
-    logger.warning("Could not serialize attribute %s:%r to tag", key, attr)
-    return None
-
-
-def _get_long_tag(key, val):
-    return jaeger.Tag(key=key, vLong=val, vType=jaeger.TagType.LONG)
-
-
-def _get_string_tag(key, val):
-    return jaeger.Tag(key=key, vStr=val, vType=jaeger.TagType.STRING)
-
-
-def _get_bool_tag(key, val):
-    return jaeger.Tag(key=key, vBool=val, vType=jaeger.TagType.BOOL)
-
-
-class AgentClientUDP:
-    """Implement a UDP client to agent.
-
-    Args:
-        host_name: The host name of the Jaeger server.
-        port: The port of the Jaeger server.
-        max_packet_size: Maximum size of UDP packet.
-        client: Class for creating new client objects for agencies.
-    """
-
-    def __init__(
-        self,
-        host_name,
-        port,
-        max_packet_size=UDP_PACKET_MAX_LENGTH,
-        client=agent.Client,
-    ):
-        self.address = (host_name, port)
-        self.max_packet_size = max_packet_size
-        self.buffer = TTransport.TMemoryBuffer()
-        self.client = client(
-            iprot=TCompactProtocol.TCompactProtocol(trans=self.buffer)
-        )
-
-    def emit(self, batch: jaeger.Batch):
-        """
-        Args:
-            batch: Object to emit Jaeger spans.
-        """
-
-        # pylint: disable=protected-access
-        self.client._seqid = 0
-        #  truncate and reset the position of BytesIO object
-        self.buffer._buffer.truncate(0)
-        self.buffer._buffer.seek(0)
-        self.client.emitBatch(batch)
-        buff = self.buffer.getvalue()
-        if len(buff) > self.max_packet_size:
-            logger.warning(
-                "Data exceeds the max UDP packet size; size %r, max %r",
-                len(buff),
-                self.max_packet_size,
-            )
-            return
-
-        with socket.socket(socket.AF_INET, socket.SOCK_DGRAM) as udp_socket:
-            udp_socket.sendto(buff, self.address)
-
-
-class Collector:
-    """Submits collected spans to Thrift HTTP server.
-
-    Args:
-        thrift_url: URL of the Jaeger HTTP Thrift.
-        auth: Auth tuple that contains username and password for Basic Auth.
-    """
-
-    def __init__(self, thrift_url="", auth=None):
-        self.thrift_url = thrift_url
-        self.auth = auth
-        self.http_transport = THttpClient.THttpClient(
-            uri_or_host=self.thrift_url
-        )
-        self.protocol = TBinaryProtocol.TBinaryProtocol(self.http_transport)
-
-        # set basic auth header
-        if auth is not None:
-            auth_header = "{}:{}".format(*auth)
-            decoded = base64.b64encode(auth_header.encode()).decode("ascii")
-            basic_auth = dict(Authorization="Basic {}".format(decoded))
-            self.http_transport.setCustomHeaders(basic_auth)
-
-    def submit(self, batch: jaeger.Batch):
-        """Submits batches to Thrift HTTP Server through Binary Protocol.
-
-        Args:
-            batch: Object to emit Jaeger spans.
-        """
-        batch.write(self.protocol)
-        self.http_transport.flush()
-        code = self.http_transport.code
-        msg = self.http_transport.message
-        if code >= 300 or code < 200:
-            logger.error(
-                "Traces cannot be uploaded; HTTP status code: %s, message: %s",
-                code,
-                msg,
-            )